backend/requirements.txt
channels==3.0.3
channels-redis==3.3.0
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
numpy==1.21.0
psycopg2==2.9.1
